+++ dit/multivariate/binding_information.py
@@ -12,6 +12,9 @@ __all__ = ('binding_information',
 
 def binding_information(dist, rvs=None, crvs=None, rv_mode=None):
     """
+    Calculates the binding information, also known as the dual total
+    correlation.
+
     Parameters
     ----------
     dist : Distribution
@@ -97,3 +100,5 @@ def residual_entropy(dist, rvs=None, crvs=None, rv_mode=None):
 
 
 dual_total_correlation = binding_information
+
+variation_of_information = residual_entropy