+++ docs/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 
 * :release:`1.0.0 <future>`
 
+* :feature:`109` Add the CAEKL mutual information
+* :feature:`108` Add the multivariate entropy triangle.
 * :feature:`63` Add scalar operations for ScalarDistributions.
 * :feature:`96` Add the marginal utility of information.
 * :feature:`95` Add the complexity profile.